refactor: Piece Abstract class
protected 메서드 추가에 따라서 piece 패키지에 있는 것이 합당하다고 생각하여
이동
