+++ src/main/java/chess/domain/ChessBoard.java
@@ -3,6 +3,7 @@ package chess.domain;
 import chess.domain.exceptions.IllegalSourceException;
 import chess.domain.exceptions.IllegalTargetException;
 import chess.domain.exceptions.InvalidRouteException;
+import chess.domain.piece.Piece;
 import chess.dto.BoardDto;
 import chess.dto.TurnDto;
