+++ src/main/java/chess/domain/generator/PieceGenerator.java
@@ -1,6 +1,6 @@
 package chess.domain.generator;
 
-import chess.domain.Piece;
+import chess.domain.piece.Piece;
 import chess.domain.Team;
 import chess.domain.piece.*;
